[fix] 실패한 테스트 코드 모두 통과되게 수정
