+++ kttr-api/src/test/java/com/crs/kttr/reservation/ReserveTrainTicketRaceConditionTest.java
@@ -1,18 +1,27 @@
 package com.crs.kttr.reservation;
 
+import com.crs.kttr.member.exception.MemberNotFoundException;
+import com.crs.kttr.member.model.Member;
+import com.crs.kttr.product.ticket.exception.TrainTicketNotFoundException;
 import com.crs.kttr.product.ticket.model.Stock;
 import com.crs.kttr.reservation.application.TicketReservationService;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 
+import java.util.Optional;
 import java.util.concurrent.BrokenBarrierException;
 import java.util.concurrent.CyclicBarrier;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.stream.IntStream;
+
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.BDDMockito.given;
 
 @SpringBootTest
 @AutoConfigureMockMvc
@@ -32,6 +41,7 @@ public class ReserveTrainTicketRaceConditionTest {
   }
 
   @Test
+  @DisplayName("락 사용하여 기차 예약 성공")
   void reserve() throws InterruptedException, BrokenBarrierException {
     ExecutorService es = Executors.newFixedThreadPool(POOL_SIZE);
 
@@ -52,5 +62,21 @@ public class ReserveTrainTicketRaceConditionTest {
     cyclicBarrier.await();
 
     es.shutdown();
+  }
+
+  @Test
+  @DisplayName("기차 티켓 예약 시 사용자 없는 경우")
+  void notFoundMember() {
+    // when & then
+    Assertions.assertThrows(RuntimeException.class,
+      () -> service.reserve(0L, 1L));
+  }
+
+  @Test
+  @DisplayName("기차 티켓 예약 시 티켓이 없는 경우")
+  void notFoundTicket() {
+    // when & then
+    Assertions.assertThrows(RuntimeException.class,
+      () -> service.reserve(1L, 0L));
   }
 }
